move settings to wpsf
