+++ settings/settings-fields.php
@@ -1,8 +1,8 @@
 <?php
 
-add_filter( 'wpsf_register_settings_solid_dynamics', 'wpsf_tabless_settings' );
+add_filter( 'wpsf_register_settings_solid_dynamics', 'wpsf_register_settings' );
 
-function wpsf_tabless_settings( $wpsf_settings ) {
+function wpsf_register_settings( $wpsf_settings ) {
 	$wpsf_settings[] = array(
 		'section_id'            => 'elementor',
 		'section_title'         => 'Elementor',
@@ -13,14 +13,21 @@ function wpsf_tabless_settings( $wpsf_settings ) {
 				'title'   => '"Back to WordPress Editor" Button',
 				'desc'    => 'Hide the "Back to WordPress Editor" button on the post page.',
 				'type'    => 'checkbox',
-				'default' => false,
+				'default' => 0,
 			),
 			array(
 				'id'      => 'hide_hello_elementor_page_title',
 				'title'   => 'Hello Elementor Page Title',
 				'desc'    => 'Hide the page title from the Hello Elementor theme.',
 				'type'    => 'checkbox',
-				'default' => false,
+				'default' => 0,
+			),
+			array(
+				'id'      => 'wrap_content',
+				'title'   => 'Wrap Elementor Content',
+				'desc'    => 'Wrap Elementor content with `main#content`.',
+				'type'    => 'checkbox',
+				'default' => 0,
 			),
 		),
 	);
